LDE-714: moved tests for extend_cube
